+++ routes/api/v1/api.php
@@ -5,6 +5,7 @@ use Illuminate\Support\Facades\Route;
 Route::get('/ping', 'App\Http\Controllers\Api\V1\PingController@index')->name('ping_index');
 // Login
 Route::post('/login', 'App\Http\Controllers\Api\V1\AuthController@login')->name('login');
+Route::post('/refresh', 'App\Http\Controllers\Api\V1\AuthController@refresh')->name('refresh');
 // Roles
 Route::get('/roles', 'App\Http\Controllers\Api\V1\RolesController@index')->name('roles_index');
 Route::post('/roles', 'App\Http\Controllers\Api\V1\RolesController@store')->name('roles_store');